pyproject.toml
[tool.poetry]
name = "small-service"
version = "0.1.0"
description = "A small service application"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "==0.115.6"}
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.36"
uvicorn = "^0.34.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.13"
taskipy = "^1.12.2"

[tool.ruff]
line-length = 120

[tool.taskipy.tasks]
lint = "ruff check --select I --fix "
format = "ruff format"
dev = "uvicorn small_service.app:app --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"